\name{anthracis}
\alias{anthracis}
\docType{data}
\title{
Bacillus anthracis str. 'Ames Ancestor'
}
\description{
This is protein details for Bacillus anthracis str. 'Ames Ancestor' with COG annotated.
If a user does not specify cog_file argument in BOG() command, this built-in COG file will be loaded by default.
}
\usage{data(anthracis)}
\format{
	  List of two elements with names: geneID, COG
}
\source{
http://www.ncbi.nlm.nih.gov/genome/proteins/181?project_id=58083
}

\examples{
data(anthracis)
}

