\name{anthracis_adjpval}
\alias{anthracis_adjpval}
\docType{data}
\title{
anthracis_adjpval
}
\description{
This is the simulated data file of anthracis. It consists of anthracis geneID and adjusted p-value for gene expression between two conditions. 
If a user does not specify the data argument in BOG() command, this built-in data will be loaded by default.
}
\usage{data(anthracis_adjpval)}
\format{
  A data frame with 5193 observations on the following 2 variables: geneID, adj.pval
}
\examples{
data(anthracis_adjpval)
}
