\name{BOG-package}
\alias{BOG-package}
\alias{BOG-package}
\docType{package}
\title{
BOG is a package for identifying differentially regulated genes in the light of gene functions.
}
\description{
It is designed to identify COG (Cluster of Orthologous Groups) that are either over or under-represented among genes that show differential expression under two different conditions using statistical analysis and graphical representations. Particular statistical analysis includes the hypergeometric test, Mann-Whitney test, and Gene set enrichment analysis (GSEA) test. The results are then organized and presented graphically for easy visualization and understanding.
}
\details{
\tabular{ll}{
Package: \tab BOG\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-3-19\cr
License: \tab GPL-3\cr
}

}
\author{
Jincheol Park, Cenny Taslim, Shili Lin, The Ohio State University

Maintainer: Jincheol Park <park.1750@mbi.osu.edu>
}
\references{
Carlson, P. et al. (2009) Transcriptional Profiling of Bacillus anthracis Sterne (34F2) during Iron Starvation. PLOS ONE 4(9): e6988. Doi:10.1371/journal.pone.0006988.

Khalili, A. et al. (2009) A robust unified approach to analyzing methylation and gene expression data. Comput. Stat. Data Anal., 53, 1701-1710.

Subramian, A. et al.(2005) Gene set enrichment analysis: A knowledge-based approach for interpreting genome-wide expression profiles. PNAS vol.102 no.43, 15545-15550.	

Taslim, C. et al. (2010) DIME: R-package for Identifying Differential ChIP-seq Based on an Ensemble of Mixture Models. Bioinformatics, 27, 1569-1570.
}

