\name{brucella}
\alias{brucella}
\docType{data}
\title{
Brucella suis 1330
}
\description{
This is protein details for Brucella suis 1330 Sequence with COG annotated.
}
\usage{data(brucella)}
\format{
  List of two elements with names: geneID, COG
}
\source{
http://www.ncbi.nlm.nih.gov/genome/proteins/806?project_id=57927
}
\examples{
data(brucella)
}

