\name{coxiella}
\alias{coxiella}
\docType{data}
\title{
Coxiella burnetii RSA 493
}
\description{
This is protein details for Coxiella burnetii RSA 493 with COG annotated.
}
\usage{data(coxiella)}
\format{
 List of two elements with names: geneID, COG
}
\source{
http://www.ncbi.nlm.nih.gov/genome/proteins/543?project_id=57631
}
\examples{
data(coxiella)
}