\name{difficile}
\alias{difficile}
\docType{data}
\title{
Clostridium difficile 630
}
\description{
This is protein details for Clostridium difficile 630 with COG annotated.
}
\usage{data(difficile)}
\format{
List of two elements with names: geneID, COG
}
\source{
http://www.ncbi.nlm.nih.gov/genome/proteins/535?project_id=57679
}
\examples{
data(difficile)
}
