\name{DIMEplot}
\alias{DIMEplot}
\title{
DIMEplot
}
\description{
This function provides visualization for the best fitted mixture model and enables a user to evaluate the fitted mixure model graphically.
Inset in the plot shows a zoomed-in plot of individual components of the model. This is relevant only if data.type="data".
}
\usage{
DIMEplot(x)
}
\arguments{
  \item{x}{
This is a BOG object.
}
}
