\name{ecoli}
\alias{ecoli}
\docType{data}
\title{
Escherichia coli O157:H7 str. Sakai
}
\description{
This is protein details for Escherichia coli O157:H7 str. Sakai
}
\usage{data(ecoli)}
\format{
  List of two elements with names: geneID, COG
}
\source{
http://www.ncbi.nlm.nih.gov/genome/proteins/167?project_id=57781
}
\examples{
data(ecoli)
}
