\name{francisella}
\alias{francisella}
\docType{data}
\title{
Francisella tularensis subsp. holarctica LVS
}
\description{
This is protein details for Francisella tularensis subsp. holarctica LVS
}
\usage{data(francisella)}
\format{
  List of two elements with names: geneID, COG
}
\source{
http://www.ncbi.nlm.nih.gov/genome/proteins/511?project_id=58595
}
\examples{
data(francisella)
}
