\name{printGSEA}
\alias{printGSEA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
printGSEA
}
\description{
If gsea argument is set TRUE in BOG() command, this command print the outcome of GSEA test. 
}
\usage{
printGSEA(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
	It is a BOG object.
}
}

