\name{BOG-package}
\alias{BOG-package}
\alias{BOG-package}
\docType{package}
\title{
BOG is a package for identifying differentially regulated genes in the light of gene functions.
}
\description{
The BOG package is designed to identify COG (Cluster of Orthologous Groups) that are over represented among genes that show differential expression under two different conditions using statistical analysis and graphical representations. Particular statistical analysis includes the hypergeometric test (HG), Mann-Whitney rank sum test (RANK), and Gene set enrichment analysis test (GSEA). The results are then organized and presented tabularly and graphically for easy visualization and presentation.
}
\details{
\tabular{ll}{
Package: \tab BOG\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-2-23\cr
License: \tab GPL-3\cr
}

}
\author{
Jincheol Park (Keimyung University, South Korea), Cenny Taslim, Shili Lin (The Ohio State University, USA)

Maintainer: Jincheol Park <park.jincheol@gmail.com>
}
\references{
Carlson, P. et al. (2009) Transcriptional Profiling of Bacillus anthracis Sterne (34F2) during Iron Starvation. PLOS ONE 4(9): e6988. Doi:10.1371/journal.pone.0006988.

Khalili, A. et al. (2009) A robust unified approach to analyzing methylation and gene expression data. Comput. Stat. Data Anal., 53, 1701-1710.

Subramian, A. et al.(2005) Gene set enrichment analysis: A knowledge-based approach for interpreting genome-wide expression profiles. PNAS vol.102 no.43, 15545-15550.	

Taslim, C. et al. (2010) DIME: R-package for Identifying Differential ChIP-seq Based on an Ensemble of Mixture Models. Bioinformatics, 27, 1569-1570.
}

