\name{DIMEplot}
\alias{DIMEplot}
\title{
DIMEplot
}
\description{
It visualize the best fitted mixture model and enables a user to evaluate the fitted mixure model graphically.
Inset in the plot shows a zoomed-in plot of individual components of the model.
}
\usage{
DIMEplot(x)
}
\arguments{
  \item{x}{
This is a BOG object.
}
}
